Nova: Port Updates from Legacy
I've been working a lot on legacy and ignoring
Nova. I've now ported all the changes from
legacy into Nova, added some tests, and ran a
live deployment.

Also synced some code with legacy so they look the
same.

Note: there is still a bug I need to fix in
the workflow that is outside of the Nova provider
before this can be used for Nova deployments, but
I'm pushing this up so ya'll can see and use it.