Deployments: incremental progress and status
A deployment goes in as 'NEW'. If you issue a
'+plan' it moves to 'PLANNED' and then you can
'+deploy' it. These statuses and operations are
all supported now. This is in preparation to be
able to execute code on +deploy that creates and
stores the certs in the deployment.