UI: fix where html requested inconsistently
This seems to be an angularjs issue. Some requests
to the server did not have application/json as an
accept header and the server returned HTML which
cause ugliness. I didn't want to dig in to
angular, so I just made sure our URLs add .json
as an extension.

Thank you, API, for being awesome!