Chef-solo: 'cook' celery task
Take this over from chef-local since we want to be
able to add attributes at cook time.

By enabling this, we can now declare that all
mapped values will be routed to and made available
for the cook task. This will allow register and
prep to proceed unimpeded and will simplify
workflows.