Deployments: fix constraint parsing bug
Constraints should be a list. That's how they are
defined in DSL.md and most examples, blueprints,
and tests.

Places where this is not true:
- in one unused app.yaml sample which this commit
  fixes.
- in the component constraints of our blueprints
  for the wordpress path values.

To maintain backwards compatibility, if a dict is
detected, it is converted to a list. This will let
the wordpress blueprints still work. This has been
marked as an item to fix and will be logged
whenever it is encountered.