Chef-solo: detect if 'relation' is in map
The 'source-key' in the relation tells us which
'requires' this relation is connected to.

Now we have the necessary relation information
being provided by the updated planning process so
we can tell if a blueprint relation is targeting a
requirement in the map file and produce the
necessary tasks to collect that data.