Fix Option Parsing Bug
Consider this:

    my_option:
      default: 'thedefaultwidgetvaluegoeshere'
      constrains:
      - type: blah
        service: foo
        setting: fa
        attribute: widget
Here's an example run that would test this and ultimately break, assuming value was the default leading into this.

value = 'thedefaultwidgetvaluegoeshere'
attribute = 'widget'
if attribute in value:
    result = value[attribute]
Traceback (most recent call last):
File "", line 2, in
TypeError: string indices must be integers, not str

In other words, the "in" facility works with string objects.
Conflicts:
	tests/test_deployments.py