Release Workflow lock at end of run_workflow
Because an executed Workflow would maintain a lock across multiple runs
it broke the ability to manually rerun a Task. In order to allow manual
task retries, release the Workflow lock as soon as run_workflow
completes.

This creates a different problem with editing running Workflows. For
now, if a Workflow is edited  and then saved while run_workflow is
executing the user will get an error because the Workflow is locked. If
they are lucky and save the Workflow in between executions of
run_workflow the Workflow will save successfully.