Chain callbacks that verify blueprint repos instead of sending them all at once
When angular makes an $http request, it sends the request directly to
the browser.  Depending on your browser, there is a max number of
requests that can be handled at a given time.  If the number of requests
exceeds that maximum, the subsequent requests are put into a queue. If a
user selected a blueprint on the page, resulting in a new $http request,
this new request was appended to the end of the queue; therefore, the
user had to wait for all the blueprint repos to be verified before
seeing his/her selection. By chaining the callbacks, the users $http
request can be handled before all the blueprint verifications come back.