Fix reset warning button. Fix modal links
The links that open modals had hrefs attributes, so the default
behavior was not being prevented - we were linking to those anchors
rather than opening the modals.  If we remove the href attribute,
angular will preventDefault on a links automagically.  Added a styling
to ensure links without href still have a pointer cursor