Patch angular to support duplicate url params
See https://github.com/angular/angular.js/pull/2522

Angular currently does not support passing in multiple param values of
the same name - it simply takes the last given value.  For example:
?my_param=1&my_param=2 is overwritten to become ?my_params=2

There is currently a PR open that resolves the issue and seems likely to
be pulled in.  Until then, we are patching angular to support url params
that share the same key

Conflicts:
	rook/static/index.html