Fixes #1103
The create operation was not taking into consideration if the deployment for which the operation was being created was a simulated deployment or not -  due to this, task would use only the DB driver for actual deployments, instead of the one for simulations.

This would cause the creation of the Delete Deployment workflow to fail for simulated deployments.

<Vineeth>