Ensure D3 node data is sorted
As we continue to receive data for the workflow tree, we need to ensure
that we feed the data to D3 in the same order. We only draw the nodes on
the page once (we only enter() once); however, we continually update the
underlying data which needs to be presented to D3 in the same order,
otherwise the drawn nodes will not correspond to the same underlying
data.