Add shims dir under tests
The concept: for external dependencies (MongoBox, SQLAlchemy, etc.), use
a shim to put a layer of abstraction between the external dependency and
our tests. This way, as the external dependency changes, we can limit
the impact of those changes to the shim, can re-use the shim in any test
and keep from having to make similar changes in many test cases.