Add ui-codemirror and ui-utils files
It looks like the old angular-ui build materials on github are
deprecated.  The resources in angular-ui are broken out into more
modular components to be added as desired.  ui-utils contains the builk
of what was angular-ui; adding ui-codemirror since we use it as well as
is not including in ui-utils