Updated tox file to prevent hard link usage
Vagrant uses a local mount in virtualbox that does not support hard
links. When tox is used it, by default, will build a dist package and
install from that. This tells tox to use the develop method instead and
to skip dist package installs. This allows vagrant and local tests to
work correctly.