fix(blueprint-cache):make create cache more atomic
Any call to create a new blueprint cache for
the given repo will delete the created directory if *any*
of the procedures ((init+pull || clone) + checkout) fail.

The check for an existing blueprint cache needs to
also check for a .git in case the previous clone failed.
The previous incorrect assumption was that the blueprint
cache directory's existence alone indicated that a good
cache was present.

Partial-Fix #192