feat(add node offline)
When adding a node to an existing deployment it is sometimes useful to
be able to bring it up in an "offline" mode. This mode will add the node
to a load balancer but instruct the load balancer to keep the node
offline (not accepting traffic).

This is accomplished through the API using +add-node and passing
"status": "offline" in the body JSON.