Copy a component instead of reference
When a service uses the same copy, we need to be sure we provide a
deepcopy instead of reference because the component has been used in
memory and has a solution for a host already and the new ones don't look
for a host.

This modifies the extensibleclass to work with copy and deepcopy
correctly.