Move gem installs out of app setup
`bundle install` takes a long time to install due to big gems like
Ridley. Moving these outside of the rest of the app install makes it
faster to rebuild containers when using cache since the gems do not have
to be reinstalled each time (as long as they haven't changed).