feat(github): added anonymous github manager
This manager mirrors a lot of GitHubManager but rips out a lot of functionality that would be irrelevant to dealing with known anonymous users such as previews and tenants. Looking forward there can likely be a decent amount of refactoring done to join the two of these to be a bit more DRY.

We use a different cache for the anonymous blueprints so as to never mix what is available for authenticated users and what is meant exclusively for anonymous users.