feat(server): add anonymous blueprint manager to router
During the process of loading up managers we now check to see if anonymous paths are explicitly disabled. If they are not we are hooking up the appropriate anonymous resource in bottle and allowing anonymous access to /anonymous routes. An instantiated AnonymousGitHubManager is also passed into our blueprints router which will in turn add the appropriate routes if one is found.