fix(contrib): add stable version of `urlparse`
The `urlparse` module from Python < 2.7.4 has problems properly
splitting/parsing URLs, specifically the query.

According to section 3.3 of rfc3986, the `path` is terminated by the
first ?, #, or end of URL. The problem we have here specifically is that
the ? is being included in the path, along with the first query param
pair. Note: This only happens when there is a single query param; if
there are 2 or more, urlparse/urlsplit function correctly.

To work around this, we have simply stolen a fixed version of urllib
from Python 2.7.6.