fix(oslo.config): now requires six 1.9.0
oslo.config's downstream dependencies were updated on 9 March, 2015.
Since oslo.config does not pin downstream dependencies to a specific
version, newer versions are being pulled in that specify a version of
six >= 1.9.0. Since we were pinning six to 1.8.0, pip was not able to
resolve the conflict.

Remove six from requirements.txt to resolve the conflict.