feat(keypairs): store keypairs in deployment
We’ve been storing key pairs as a resource, but we
need all deployments to have keys that we can use
for authentication and managing resources. This
change mirrors key pairs to `key-pairs` at the
root of the deployments, but maintains
compatibility with existing code by also storing
key pairs as a resource.

The resource key has been moved into a constant
as well.