fix(rax.base): check if tenant is a digit first
Tenants in tests are not digits. And tenants are
always strings (not ints). We also plan to support
tenants starting with hybrid: and ‘D’ for
dedicated accounts. In fact, that was why we
intentionally used alphanumeric tenants in our
tests. This makes our code that checks for UK
tenants more defensive.