feat(tests): pylint allow for _module_vars
We debated the confusing aspect of CONSTS that
are mutable (module-level variables that we use
for caching, for example). We decided to use
regular variable names instead of ALL_CAPS which,
by convention, indicates a constant. To
disambiguate module-level variables from local
variables, we decided to require an _ at the
beginning.

Pylint would still detect if a local variable
overrides such a module-level variable and raise a
W0621 Redefining name from outer scope warning.