fix(bootstrap): Add multiple retries in response to passwordsafe's random 403's
This is to resolve issues we often see with passwordsafe throwing 403s to valid auth tokens. The token we are supplying is know valid so we want to resend the request again. We've found that PS will often continue to fail for a short period of time so a sleep was introduced. This makes the retries more effective by allowing PS to return to normal behavior and allow us access.