fix(chef-server): pass deployment keys to kitchen
This code was copied from chef-solo, but one part
seems to be missing. The deployment keys were not
being passed in, so a key was generated to write
into the kitchen’s private.pem file, but that key
is never placed on the servers. This fixes this by
passing in the keys to the kitchen creation so
that a key is not generated, but the right key is
written that would access the server.