fix(rax.compute.provider): obviate region errors
When you attempt to deploy a blueprint, using pawn, and select a region
which is incompatible with the specified account, the error message you
get isn't very useful:

Error: 500. Could not find a provider that can create a component in the
server service - Internal Server Error

I had to debug things until I figured that this error occurs when you
select a region that is incompatible with your account (e.g., selecting
LON with a non-UK account).

This change fixes that by throwing a more useful error message about the
region. With this, pawn users will at least know there is something
wrong with the region, and they'll try another one.

For reference, I found this error by testing the blueprint at
https://github.com/checkmate/helloworld.