fix(database replica)
When searching for the master database, sometime the database instance
was found after the database itself, causing the `master-db-id` to be
improperly set as the database instance. `master-db-id` should always
point to the master database, _not_ the master database _instance_.

Add additional check to `if` condition so that only the database itself
is considered when setting the id for `master-db-id`.