BLUEPRINT_CACHE_EXPIRE no longer used
The validity/freshness of a cache for a given blueprint
repository is now simply based on whether that repository
is *actually* up-to-date by verifying it every time. We
enabled this verificaiton to happen efficiently by using
`git ls-remote` and comparing our local (cached) revisions
to the revisions (commit hashes) on the remote. If the
ls-remote query tells us we are up-to-date, we don't need
to fetch/checkout. If the ls-remote query shows us that we
don't have the latest revision for the specified ref
(note: the specified ref defaults to 'master') only then
do we perform the fetch/checkout.